[hms] Avoid blocking table alterations that are not relevant to Kudu
This patch adjust the KuduMetastorePlugin to prevent blocking table
alterations that are not relevant to Kudu metadata which needs to be
synchronized. Additionally it prevents calling the Kudu master to check
if HMS sync is enabled when the metadata changes are not relevant to
Kudu’s HMS synchronization.

Change-Id: I2517d891ab7168164700bb3ae642c49bde54b9db
Reviewed-on: http://gerrit.cloudera.org:8080/17531
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
(cherry picked from commit 3667428d9071cbcf6825b07983529eddffc86dd5)
Reviewed-on: http://gerrit.cloudera.org:8080/17537
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>