[test] make new scenario in security-itest more robust
This patch makes the newly introduced test scenario
SecurityITest.TxnSmokeWithDifferentUserTypes more robust and stable.
With awaiting for the first transaction to finalize, now there is no
risk for aborting the second transaction in case of a race.

Before this patch, if it took a long time to finalize aborting the
former transaction, the deadlock prevention logic would automatically
abort the latter one, failing the test scenario.

Before this fix, 1/4 (one if four) of runs failed if running with
--stress_cpu_threads=16.  With this fix, I didn't see a failure
in hundreds of runs.

Change-Id: Ie99ad41a7a161f9046ba894c072eb92150f0e0e2
Reviewed-on: http://gerrit.cloudera.org:8080/17559
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>