[docker] Add option to force the latest tag
This patch adds a `--force-latest` that allows users to force
using the simple `latest` tag even when the default base OS
is not used. This is useful because the latest tag is often
used to retrieve the last relevant build and is commonly used
when a user just wants to test or use their latest build.

For example if a user wants to test the quickstart with CentOS 7
they can build with `--force-latest` and run normally as opposed
to manually editing the quickstart.yml file.

This is also useful for CI/CD pipelines that always want to publish
the last build image as `latest`.

Change-Id: I13ff38683d3f923577f8728fdf30a4b3e09bc943
Reviewed-on: http://gerrit.cloudera.org:8080/17574
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>