docs: use PNG instead of SVG for image
The SVG resulted in a failed RAT check for the release, so this patch
updates to using a PNG, which is more common in our codebase anyway. It
also moves it into the /images folder in /docs.

Rendered version:
https://github.com/andrwng/kudu/blob/txns_docs/docs/design-docs/transactions.adoc

Change-Id: I6dff54eb4853e116b8da2ef2c604fe3de588bf5f
Reviewed-on: http://gerrit.cloudera.org:8080/17578
Tested-by: Kudu Jenkins
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>
(cherry picked from commit 38f037e6622a4b3916418ed01186ec97ef136087)
Reviewed-on: http://gerrit.cloudera.org:8080/17579
Reviewed-by: Andrew Wong <awong@cloudera.com>