[clock] change clock source selection for 'auto'
This patch changes the logic to select particular time source when
the pseudo-source 'auto' used for --time_source.

Before this patch, the 'system' time source would be auto-selected if
a Kudu server with --time_source=auto is run in an environment where the
instance detector isn't aware of a dedicated NTP server (those are
usually available via the host-only interface, at least for EC2 and
GCE instances).

After this patch, the 'builtin' time source would be auto-selected
if a Kudu server runs with --time_source=auto in environment where the
instance detector isn't aware of dedicated NTP servers AND the
--builtin_ntp_servers flag is set to a valid value.  Otherwise, if
--builtin_ntp_servers flag is set to an empty/invalid value, 'system'
becomes the auto-selected time source for platforms supporting
get_ntptime() API, otherwise the catch-all case is used to auto-select
'system_unsync' as the time source.

I also added a new test scenario to cover the updated functionality.

Change-Id: I72cb54c488b2aa2c73acd6e6e5f6e50dd5811175
Reviewed-on: http://gerrit.cloudera.org:8080/17582
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>