[client] actionable error message for KuduSession::Flush()
This patch makes the error message for Status::IOError() returned by
KuduSession::Flush() more actionable, citing that details are available
via call to KuduSession::GetPendingErrors().  Yes, that's documented
in-line in client.h and in the client C++ API docs [1], but who reads
those?

Prior to this patch, the error message was:
  Some errors occurred

With this patch, the error message now is:
  failed to flush data: error details are available
  via KuduSession::GetPendingErrors()

In addition, I switched the code in batcher.{h,cc} from Atomic
(util/atomic.h and gutil/atomicops.h) to std::atomic and removed the
Batcher::MarkHadErrors() utility method.

[1] https://kudu.apache.org/cpp-client-api/classkudu_1_1client_1_1KuduSession.html

Change-Id: I69e1613d703bb689641186c4282fe8b3e821d323
Reviewed-on: http://gerrit.cloudera.org:8080/17673
Tested-by: Kudu Jenkins
Reviewed-by: Bankim Bhavsar <bankim@cloudera.com>