[webserver] tag a few flags as sensitive
This patch marks a few flags of the embedded webserver as 'sensitive',
forcing the redaction of their values while logging those in the logfile
or showing them in the UI served by the embedded webserver.  Yes, that's
rather "security by obscurity" practice, but it could help in preventing
unintentional leaks in the absence of proper FS access restrictions
and in avoiding false-positives reported by various automated security
scanners.
  * --webserver_private_key_file
      this file may contain a private key file in non-encrypted form
  * --webserver_private_key_password_cmd
      the value of this flag may contain a password for the private key
      file in a plain form or other sensitive info provided as arguments
      for the command
  * --webserver_password_file
      this file contains user names and encrypted/hashed passwords

Change-Id: I187c6e8118a3e118299da1e3f7321f48c5eb9e05
Reviewed-on: http://gerrit.cloudera.org:8080/17679
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins