[catalog_manager] introduce --min_num_replicas flag
This patch introduces a flag for kudu-master to enforce the minimum
number of replicas for newly created tables in a cluster.  Unless
overridden, --min_num_replicas is set to 1 by default.

For example, setting --min_num_replicas=3 enforces every new table to
have at least 3 replicas for each of its tablets, so there cannot be
a data loss when a single tablet server in the cluster fails
irrecoverably.

I also added validators for the related flags to keep things more
consistent and report misconfiguration as early as possible.

This patch contains a few test scenarios to cover the newly introduced
functionality.

Change-Id: I86191fcdc1b4ed6670f33ba7176d28dbd1df541f
Reviewed-on: http://gerrit.cloudera.org:8080/17684
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>