KUDU-2671 remove semantically duplicate range_hash_schemas
This patch removes the 'range_hash_schemas' field from
CreateTableRequestPB because the field was semantically duplicating the
'range_hash_schemas' sub-field of the already existing
'partition_schema' field.  This change doesn't break any compatibility
because the corresponding client side which uses the removed field
hasn't yet been released.

In addition, this patch fixes an invalid memory access condition
(sometimes leading to SIGSEGV) in PartitionSchema::EncodeRangeBounds()
if the number of per range hash schemas is less than the number
of range bounds.

With the removal of the semantically duplicate field, the check for
the validity of per-range hash bucket ranges is now effective.  This
patch adds a new test scenario to verify that the validation is now
in place and to catch regressions in the future.

I also updated the corresponding code in the C++ client and tests.

This is a follow-up to 23ab89db1 and 586b79132.

Change-Id: Icde3d0b0870fd3a3941fcc91602993ae7ad46266
Reviewed-on: http://gerrit.cloudera.org:8080/17694
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>