[partition] RangeHashSchema --> PerRangeHashBucketSchemas
This patch simply renames the PartitionSchema::RangeHashSchema type
into PartitionSchema::PerRangeHashBucketSchemas, updating the existing
typedef for std::vector<HashBucketSchemas>.  I found it's easier to read
and comprehend the code with this change because the PB-based
PartitionSchemaPB::PerRangeHashBucketSchemasPB structure is very close
to what PartitionSchema::PerRangeHashBucketSchemas is.

In addition, this patch introduces a RangesWithHashSchemas typedef for
vector<RangeWithHashSchemas> and updates the signature of the
PartitionSchema::SplitRangeBounds() method.

I also updated the code in master-test and table_locations-itest to
introduce similar typedefs.  I found some more room for refactoring
in those two test suites (e.g., there are similar data structures
in both), but I decided not to address that right away to avoid
bloating of this patch.

This patch does not contain any functional changes.

Change-Id: Ic64a956a87c3d66cb2a14f4a3da7f57948518f17
Reviewed-on: http://gerrit.cloudera.org:8080/17724
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>