KUDU-1921 Add ability to require auth/encryption to C++ client
Kudu servers support requiring authentication and encryption to be
enabled, and clients prefer connecting in a secure way, but if a server
doesn't support authentication and/or encryption, the client will
silently connect insecurely, which can lead to a downgrade attack.

With this patch, clients can require authentication and encryption to be
set using the client API, where if such an attack is attempted, the
client will fail to connect to the cluster.

Change-Id: Ia3e800eb7c4e2f8787f0adf1f040d47358d29320
Reviewed-on: http://gerrit.cloudera.org:8080/17731
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>