KUDU-1959 - Implement server startup progress page for tablet and master servers
This patch implements adjusting the server startup sequence to start the
web server and load the startup page which shows the progress of the startup.
This progress is broken down into the below steps:

Initializing
Reading Filesystem
  Reading instance metadata files
  Opening container files
Bootstrapping and opening the tablets
Starting RPC server

Sample screenshot  of the page can be found here:
https://i.imgur.com/J4NJD5i.png

Each of the steps above has a progress status of either 0 or 100 except
for the step “Bootstrapping and opening the tablets” which tracks the
processing of each of the tablet and the step “Opening container files”.
For log block manager this step shows the progress of opening the log
block containers. For file block manager, this is renamed to
“Reporting Filesystem” and has only 0 or 100 progress status. All of
the steps have time elapsed presented in seconds.

For the master startup page, the step “Bootstrapping and opening the
tablets” is replaced by “Initializing master catalog”.

Along with the above startup page the web server will also display
the pages - /home, /config, /logs, /mem-trackers, /memz, /threadz, /varz.
During the initialization since we expect the user's primary reason to
open the web UI to be to check the startup progress, /home  will be
redirected to /startup until the startup is completed. The usual homepage
contents and the other remaining web pages will be available once the rpc
server is started due to dependency. Manually validated the startup when
security is enabled and filesystem is not present.

The footer of the WebUI has the UUID by default. In case of new
instances, since we are starting the Webserver before the UUID is
generated the footer doesn’t contain UUID but it is set later when
the rpc server is started.

The next steps for this patch include writing tests for the startup
page. So far all the testing has been manual.

Change-Id: I1db1fcf16261d4ced1b3657a697766a5335271b4
Reviewed-on: http://gerrit.cloudera.org:8080/17730
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>