[test] deflake ToolTest.TableScanCustomBatchSize
There seems to be a bug when running loadgen with AUTO_FLUSH_BACKGROUND
mode, wherein background write flushes aren't waited for before the
process is terminated, resulting in a leak report. I'll address the
underlying issue separately, but for now, this patch adjusts the test to
use fewer rows so no leaks are reported.

The flaky test dashboard reports the test as ~10% flaky. With this
patch, I ran it in ASAN and it passed 100/100 times.

Change-Id: I0740de18a3425cfd47aefe354b6a9fe05aa291cf
Reviewed-on: http://gerrit.cloudera.org:8080/17747
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>