KUDU-2671 key encoding for custom hash bucket schemas
This patch updates the PartitionSchema class to properly encode range
keys in the presence of range partitions with custom hash bucket schemas
in a Kudu table.  I added a few new test scenarios and updated the
existing ones accordingly to make it possible to send write operations
into appropriate tablets.

In addition, I'm sneaking in minor style-related updates in
  * src/kudu/tablet/tablet.cc
  * src/kudu/master/catalog_manager.cc

As one can see, the newly added method GetHashBucketSchemasForKey() is
used in PartitionSchema also in case when Partition object is at hand.
A follow-up changelist is going to change that, so there will be no need
to perform a dictionary lookup once the Partition class contains
information about its hash bucket schema.

This is a follow-up to 586b7913258df2d0ee75470ddfb2b88d472ba235.

Change-Id: I81aa1c10998d88a1bd5314fc3132037b1c8bbe4b
Reviewed-on: http://gerrit.cloudera.org:8080/17769
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>