[common] more generic API for IN list predicate pruning
While working on KUDU-2671, I found that the exposed internals of the
PartitionSchema class doesn't allow for updating the implementation of
the partition-related code to include per-range custom hash bucket
schemas in a consistent manner.

This patch introduces a bit more generic interface for pruning values of
IN list predicates by adding a new PartitionMayContainRow() method and
removes the following methods from the public API of the PartitionSchema
class:
  * HashPartitionContainsRow()
  * RangePartitionContainsRow()
  * IsColumnSingleRangeSchema()
  * TryGetSingleColumnHashPartitionIndex()

I also added one extra test scenario and updated existing ones to
increase readability of the assertion messages if they are triggered.

This is a follow-up to 6a7cadc7e and 83b8caf4f.

Change-Id: I2e2390cc4747864fdac71656dd7125ac3b15bf9d
Reviewed-on: http://gerrit.cloudera.org:8080/17764
Tested-by: Kudu Jenkins
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>