[partition] update naming of related entities
With this patch, the naming for the partition-related entities
in the C++ code has changed:
  * renamed HashBucketSchema into HashDimension
  * renamed HashBucketSchemas into HashSchema, so HashSchema is now
      a typedef for std::vector<HashDimension>
  * partitioning schema --> partition schema
  * hash bucket schema --> hash schema
  * hash partitioning schema --> hash schema
  * range partition schema --> range schema
  * hash_bucket_schemas_ --> hash_schema_
  * removed the PerRangeHashBucketSchemas typedef

The idea is to represent the hash-related component of the partition
schema as a set of hash bucket dimensions.  Before, it was called
'hash bucket schemas' to be 1-to-1 mapping with the
PartitionSchemaPB::HashBucketSchemaPB protobuf message defined in
common.proto.  I found it a bit awkward calling a single dimension of
hash bucketing 'schema' because it's semantically confusing, since
'schema' usually means the whole set of rules, not just a part of those.
It would be nice to update the proto-related names as well, but doing so
would break backwards compatibility.

I also updated the in-line documentation for the partition-related
entities in the PartitionSchema's class-wide comment and added TODOs
where more updates are required to accommodate the recent changes
introduced in the context of KUDU-2671.

Change-Id: I6a858e97090930b21e9c767dac2f5cc8b9816033
Reviewed-on: http://gerrit.cloudera.org:8080/17775
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>