KUDU-2671 allow a range to have empty hash schema
With this patch, the semantics of an empty hash schema for a range
changes.  Now an empty per-range hash schema means no hash bucketing for
the range.  Prior to this patch, an empty hash schema for a range meant
using the table-wide hash schema.

The new semantics are better because:
  * they allow for having ranges with no hash bucketing even if there is
    a non-trivial hash bucketing at the table-wide level
  * they are less surprising to a user of the client API

This patch updates several test cases to account for the change and
adds a new test case to cover the new functionality.

Change-Id: Ia43df69ecd7040e285e098fde49d84a7a00d1fbb
Reviewed-on: http://gerrit.cloudera.org:8080/17825
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>