KUDU-2671 test for unbounded range with custom empty hash schema
This patch adds a couple of new test scenarios to cover a case
of a table with a range partition without the upper bound,
where the range partition has a custom empty hash schema in the presence
of non-empty table-wide hash schema.  This is to make sure:
  * it's possible to create such a table
  * the range and partitions keys for the table are encoded as expected
  * it's possible to insert rows into the table

This is a follow-up to 03451904a20123ca27eaa4e9773b94c0532fd342.

Change-Id: I2198714a8025ae4b132afc4cd4700de18df226cb
Reviewed-on: http://gerrit.cloudera.org:8080/17836
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>