[common] small re-factoring on PartitionPruner
This changelist simplifies the signature of the
PartitionPruner::ConstructPartitionKeyRanges() method: the in-out
parameter is removed and turned into the return value.

I also moved RangeBoundsAndPartitionKeyRanges, RangeBounds, and
PartitionKeyRange structures into the private section of the
PartitionPruner class and removed default constructor declarations
for them.  There is no point of declaring default constructors for
structures: the compiler creates them automatically.

This patch doesn't contain any functional changes.

Change-Id: I38c5c4809f251c8efff0de584bbbbbc03e9d952e
Reviewed-on: http://gerrit.cloudera.org:8080/17854
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Mahesh Reddy <mreddy@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>